Reorg client virus index code to display a detail view
